modelpilot: web dashboard + conversation-continuation model
- dashboard.py: /modelpilot/dashboard (server-rendered HTML, inline SVG,
  no JS/CDN deps) and /modelpilot/stats JSON — headline cards, cumulative
  savings curve, daily model-mix migration chart, category drill-down,
  RCT verified-saving panel with bootstrap CI, escalation netting
- continuation.py: expected-remaining-turns via mean residual life over
  observed session lengths, replacing the flat 5-turn guess in the
  cache-switch economics; falls back to default until 50 sessions observed
- ledger: session_key column + turns_so_far/session_lengths/daily_series/
  daily_model_mix queries
- gateway: wires continuation estimate into decide(), records session_key,
  mounts dashboard routes ahead of the catch-all passthrough

41 tests passing; dashboard smoke-tested through the gateway with a
seeded 14-day demo ledger (RCT panel live at 26% verified saving).

+++ modelpilot/router.py
@@ -137,7 +137,7 @@ def classify(features: dict) -> tuple[str, int, float, str]:
 
 def recommend(
     body: dict,
-    expected_remaining_turns: int = 5,
+    expected_remaining_turns: float = 5.0,
     classifier=classify,
 ) -> Recommendation:
     """Full routing decision for one request body."""
@@ -176,8 +176,8 @@ def recommend(
         original_model,
         candidate,
         cached_prefix_tokens=cached_prefix,
-        expected_remaining_input_tokens=per_turn_in * expected_remaining_turns,
-        expected_remaining_output_tokens=per_turn_out * expected_remaining_turns,
+        expected_remaining_input_tokens=int(per_turn_in * expected_remaining_turns),
+        expected_remaining_output_tokens=int(per_turn_out * expected_remaining_turns),
     )
     if benefit is not None and benefit <= 0:
         return stay(
